requirements.txt
pandas
numpy
scikit_learn==0.22.2.post1
torch>=1.5.0
mysql-connector-python-rf
ConfigSpace
smac
openbox
hyperopt
botorch
psycopg2
psycopg2-binary
matplotlib
pandas
#tensorflow
psutil
pyDOE
#kdtree
scipy
sklearn
cython
scikit-optimize
#smac
#botorch
#gpytorch
aliyunsdkcore
aliyun-python-sdk-rds
dynaconf
